pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qportfolio"
version = "0.1.0"
description = "Quantum-ready crypto portfolio optimization scaffold"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "OpenAI"}]
dependencies = [
  "pydantic>=2.6",
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
  "mypy>=1.10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
